Support WASM with GHC 9.8
Disabled dependencies are mostly due to `network` unless otherwise stated.

Many of the workarounds here may also apply to the JS backend. But I'm not testing with that at the moment, since when I last did I hit a lot of run-time errors. So we'll cross that bridge when we come to it. The WASM backend seems a lot more robust.

Note that I'm still using GHC 9.6 for native builds (largely because there somehow isn't yet a Cabal release which strictly supports it), and thus haven't tested there. But 9.8 is required for WASM, at least for `bitvec` which is a transitive dependency of `Rasterific`.