Allow building with GHC 9.10
Note that we specify minimal `allow-newer`s. The previous approach caused confusion a few times, such as leading to an older `lens` being allowed and obscuring that `Chart` was the real problem. In changing this, we also move them to the project file. As for the Cabal bug mentioned in the removed comment, it can easily be worked around by using a dummy package. I may open a bug report anyway but it's likely to be seen as low-priority given the current state of `install --lib`.