Allow cross-compilation with GHC 9.10
Involves pointing to new force-pushed patches for Servant (based on last release) and SBV (based on HEAD, for no particular reason).

It's not totally clear how I've gotten away without specifying `PKG_CONFIG_PATH` in the past, but it seems to be necessary now. Or at least, it's the most reliable way to get libraries like `zlib` and `evdev` to compile.