Fix Wasm build
We haven't been keeping this up-to-date for a year or so. But we're now again testing with the TH-supporting GHC 9.10 from GHCup.

Note that we've disabled some packages which fail with obscure errors regarding dynamic linking and shared interface files. Given the current experimental nature of the Wasm backend, it seems best not to get too bogged down in investigating these failures.