Add Chart example
There's a lot here that I otherwise forget every time I reach for the `Chart` library, including to some extent the basics of setting up axes etc. This is particularly frustrating as it's the kind of library one often wants to just quickly reach for in order to try something.

The `diagrams` backend is too slow for live reloading. See https://github.com/timbod7/haskell-chart/issues/105.

Unicode support is poor (in both backends), hence the use of lambda syntax rather than `↦`.

We use `+ 0.001` as a hack to prevent discontinuities for some common functions.