Initial commit of home dashboard
Code is fairly messy. CSS approach is a bit of an experiment with mostly purely structural selectors, and may not be sustainable.

This is designed to run constantly on my TV. Therefore the visuals have been designed to look good on a particular 1080p screen, with responsiveness not being very important. It may turn out to by useful on phone etc. as well, but that's not a goal for now.