Implement weather location params API in terms of explicit sum and product
The idea here is that by generalising the notion of sums and products for parameters and paths etc. we avoid needing so many bespoke interpreter-specific instances of classes like `miso`'s `ToFetch` and `servant-client`'s `HasClient`. In a sense, this approach is more declarative, with more information being given in the API definition itself.

Note that the function generated by `Product` has to be uncurried so that we can use these combinators recursively.