Use simpler type families
The previous approach was largely an attempt to see whether we could use some simple library of type families, but there seems to be nothing simple and standard.

This is pretty readable now, anyway. The fact that the link between `a'` and `Arg (ToFetch (a :> api))` is a bit indirect is weird, but I don't think we can do better, because there are significant limitations around using type families with quantified constraints.