Bump Miso
This was necessary on my laptop since the previously-pinned version wouldn't build with Wasm, with some weird errors like `unknown type name 'HsIORef'`.

There have been a few changes to the components API. These are handled entirely within `MisoCounter.hs` and `HomeDashboard.hs`.

Other changes are due to Miso dropping it's internal Servant support in favour of `servant-client-js`, obsoleting my branch which we were previously using. This is mostly fairly straightforward to adapt to. It's not quite clear exactly why this change has forced the use of HTTPS for the Spotify URL, but it's no bad thing, and appears to be something to do with CORS.

The index-state bump brings in a newer `jsaddle-wasm`, required by the newer Miso. We bump no more than a day more than necessary, in order to avoid a major rebuild.