Use better Mandelbrot colour function
This is fairly close to the original Python behaviour, but IMO slightly better. There are some arbitrary values in the inputs to `hsv`: the hue ranges over 290 rather than 360 so that high values are recognisably different from low ones, rather than both being red; and the value multiplier of 5 ensures that only the initial fifth of inputs are coloured darker.