Move to new OpenWeatherMap API (One Call API 3.0)
This was initially motivated by wanting access to UV index data, though I'm sure we'd have wanted to move eventually anyway for one reason or another, or being forced to.

Note that the free API limit is a lot lower, but we can still afford 3 concurrent clients, as mentioned in a source comment. Calls over the limit are charged at £0.0012, but by limiting to 1000 in the account settings at https://home.openweathermap.org/subscriptions, it's seemingly possible to avoid charges entirely.