Common up weather record variations
This is pretty neat for type-level programming. The need for `SBoolI (f == Current)` and the rest in the `FromJSON` signature is a bit unfortunate, since we should in principle by able to start more directly that the result of `==` is a boolean, but no better way is obvious.

Ideally, we'd have the unit fields not even exist, so that this is more analogous to the similar approach for sums using `Void`. We might be able to hide its `HasField` instances by using some sort of additional wrapper types and type families, but it's unlikely to be worth the faff. Same goes for various other improvments, such as better inference in `parseJSON`.