Prevent movement after hard drop
A side effect of this is that row elimination on hard drop is immediate, with no frame showing the completed row. This is weird, but arguably more consistent with our overall simple approach, and kind of cool. If we want an option for animations, that should be its own thing.