Use colour brightness to indicate layer
Note that `G.mkGraph` is strictly more flexible than the one in `diagrams-graphviz`. It it allows us to use proper labels separately from IDs, and doesn't require that we have the nodes' _labels_ in the edge list, which would now be problematic due to the layer data.

The initial singleton map is a bit ugly, but works fine. There's no obvious principled way to detect cycles without major changes.